requirements.txt
aiohttp~=3.6.2
discord~=1.0.1
PyYAML~=5.3.1
Pillow~=7.2.0
requests